web-scraper/requirements.txt
attrs==23.1.0
certifi==2023.11.17
click==8.1.7
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
Pillow==10.1.0
PySocks==1.7.1
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
